Crate dock_core

Re-exports

pub use modules::accumulator;

pub use modules::anchor;

pub use modules::attest;

pub use modules::blob;

pub use modules::did;

pub use modules::master;

pub use modules::offchain_signatures;

pub use modules::revoke;

pub use modules::status_list_credential;

pub use modules::trust_registry;

Modules

common

modules

runtime_api

util

Macros

def_state_change

Defines StateChange using supplied actions.

deposit_indexed_event

Deposits an event indexed over the supplied fields.

field_accessor

Implements field accessor based on input using supplied self.

hex_debug

Implements Debug trait printing hex representation of the underlying value.

impl_action

Implements Action trait with supplied params for the given ident(s).

impl_action_with_nonce

Implements Action and ActionWithNonce traits with supplied params for the given ident(s).

impl_authorize_target

impl_bits_conversion

Implements bits conversion from/to the given type for the supplied bitflags identifier.

impl_encode_decode_wrapper_tests

Implements Encode/Decode wrapper tests for the supplied wrapper and type.

impl_to_state_change

Implements ToStateChange trait with supplied params for the given ident(s).

impl_tuple

Implements given trait for the tuple type.

impl_wrapper

Implements two-direction From/Into and one-direction Deref/DeferMut traits for the supplied wrapper and type.

impl_wrapper_from_type_conversion

Implements From<type> for the wrapper.

impl_wrapper_type_info

Implements type’s type info for the wrapper.

pub_for_test

Makes given ident pub only for test and uses supplied visibility if compiled otherwise.