Trait dock_core::common::authorization::AuthorizeTarget

pub trait AuthorizeTarget<T, Target, Key>where
    Target: Associated<T>,{
    // Provided method
    fn ensure_authorizes_target<A>(
        &self,
        _: &Key,
        _: &A,
        _: Option<&Target::Value>
    ) -> DispatchResult
       where A: Action<Target = Target> { ... }
}

Authorizes action performed by Self over supplied target using given key.

Provided Methods

fn ensure_authorizes_target<A>( &self, _: &Key, _: &A, _: Option<&Target::Value> ) -> DispatchResultwhere A: Action<Target = Target>,

Self can perform supplied action over target using the provided key.