Trait dock_core::common::policy::HasPolicy

pub trait HasPolicy<T: Limits>: Sized {
    // Required method
    fn policy(&self) -> &Policy<T>;

    // Provided methods
    fn execute_view<A, F, R, E>(
        self,
        f: F,
        action: A,
        proof: Vec<DidSignatureWithNonce<T>>
    ) -> Result<R, E>
       where T: Config,
             F: FnOnce(A, Self) -> Result<R, E>,
             WithNonce<T, A>: ActionWithNonce<T> + ToStateChange<T>,
             <WithNonce<T, A> as Action>::Target: StorageRef<T>,
             E: From<PolicyExecutionError> + From<Error<T>> + From<NonceError> + From<ActionExecutionError> { ... }
    fn execute<A, F, R, E>(
        &mut self,
        f: F,
        action: A,
        proof: Vec<DidSignatureWithNonce<T>>
    ) -> Result<R, E>
       where T: Config,
             F: FnOnce(A, &mut Self) -> Result<R, E>,
             WithNonce<T, A>: ActionWithNonce<T> + ToStateChange<T>,
             <WithNonce<T, A> as Action>::Target: StorageRef<T>,
             E: From<PolicyExecutionError> + From<Error<T>> + From<NonceError> + From<ActionExecutionError> { ... }
    fn execute_removable<A, F, R, E>(
        this_opt: &mut Option<Self>,
        f: F,
        action: A,
        proof: Vec<DidSignatureWithNonce<T>>
    ) -> Result<R, E>
       where T: Config,
             F: FnOnce(A, &mut Option<Self>) -> Result<R, E>,
             WithNonce<T, A>: ActionWithNonce<T> + ToStateChange<T>,
             <WithNonce<T, A> as Action>::Target: StorageRef<T>,
             E: From<PolicyExecutionError> + From<Error<T>> + From<NonceError> + From<ActionExecutionError> { ... }
}

Denotes an entity which has an associated Policy.

Required Methods

fn policy(&self) -> &Policy<T>

Returns underlying Policy.

Provided Methods

fn execute_view<A, F, R, E>( self, f: F, action: A, proof: Vec<DidSignatureWithNonce<T>> ) -> Result<R, E>where T: Config, F: FnOnce(A, Self) -> Result<R, E>, WithNonce<T, A>: ActionWithNonce<T> + ToStateChange<T>, <WithNonce<T, A> as Action>::Target: StorageRef<T>, E: From<PolicyExecutionError> + From<Error<T>> + From<NonceError> + From<ActionExecutionError>,

Executes action over target data providing a mutable reference if all checks succeed.

Checks:

Verify that proof authorizes action according to policy.

Verify that the action is not a replayed payload by ensuring each provided controller nonce equals the last nonce plus 1.

Returns a mutable reference to the underlying data wrapped into an option if the command is authorized, otherwise returns Err.

fn execute<A, F, R, E>( &mut self, f: F, action: A, proof: Vec<DidSignatureWithNonce<T>> ) -> Result<R, E>where T: Config, F: FnOnce(A, &mut Self) -> Result<R, E>, WithNonce<T, A>: ActionWithNonce<T> + ToStateChange<T>, <WithNonce<T, A> as Action>::Target: StorageRef<T>, E: From<PolicyExecutionError> + From<Error<T>> + From<NonceError> + From<ActionExecutionError>,

Executes action over target data providing a mutable reference if all checks succeed.

Checks:

Verify that proof authorizes action according to policy.

Verify that the action is not a replayed payload by ensuring each provided controller nonce equals the last nonce plus 1.

Returns a mutable reference to the underlying data wrapped into an option if the command is authorized, otherwise returns Err.

fn execute_removable<A, F, R, E>( this_opt: &mut Option<Self>, f: F, action: A, proof: Vec<DidSignatureWithNonce<T>> ) -> Result<R, E>where T: Config, F: FnOnce(A, &mut Option<Self>) -> Result<R, E>, WithNonce<T, A>: ActionWithNonce<T> + ToStateChange<T>, <WithNonce<T, A> as Action>::Target: StorageRef<T>, E: From<PolicyExecutionError> + From<Error<T>> + From<NonceError> + From<ActionExecutionError>,

Executes action over target data providing a mutable reference if all checks succeed.

Unlike execute, this action may result in a removal of a data, if the value under option will be taken.

Checks:

Verify that proof authorizes action according to policy.

Verify that the action is not a replayed payload by ensuring each provided controller nonce equals the last nonce plus 1.

Returns a mutable reference to the underlying data wrapped into an option if the command is authorized, otherwise returns Err.