Module dock_core::modules

Modules

accumulator

anchor

Minimal proof of existence registry.

attest

This module allows DIDs to publically attests to arbirary (and arbitrarily large) RDF claimgraphs. These attestations are not stored on-chain; rather, the attester chooses a storage method by specifying an Iri.

blob

Generic immutable single-owner storage.

did

master

Simulates a multisig root account.

offchain_signatures

Module to store offchain signature keys and parameters for different signature schemes. Currently, can be either BBS, BBS+, Pointcheval-Sanders or BBDT16. For BBS, BBS+ and Pointcheval-Sanders, the public key is in group G2 but for BBDT16, it’s in group G1 and used to verify the proof of validity of MAC (and related proofs) but not the MAC itself.

revoke

status_list_credential

Storage for status list-related verifiable credentials:

trust_registry

Dock Trust Registry.