Type Definition dock_core::modules::did::StoredOnChainDidDetails

pub type StoredOnChainDidDetails<T> = WithNonce<T, OnChainDidDetails>;

Each on-chain DID is associated with a nonce that is incremented each time the DID does a write (through an extrinsic). The nonce starts from the block number when the DID was created to avoid replay attacks where an action of a DID that is removed and recreated by the same owner and same key is replayed by someone else.