Module dock_core::modules::trust_registry::types

Structs

AggregatedIssuerInfo

AggregatedTrustRegistrySchemaMetadata

TrustRegistry schema metadata.

Convener

Trust registry Convener’s DID.

ConvenerOrIssuerOrVerifier

Trust registry Convener/Issuer/Verifier’s DID.

DelegatedIssuerSchemas

Set of schemas corresponding to a delegated Issuer

DelegatedIssuers

Delegated Issuers.

DelegatedSchemaCounter

Issuer

Trust registry Issuer’s DID.

IssuerAndVerifier

Both an Issuer and a Verifier.

IssuerOrVerifier

Maybe an Issuer or a Verifier but definitely not a Convener.

IssuerSchemas

Set of schemas corresponding to an Issuer.

IssuerTrustRegistries

Set of trust registries corresponding to a issuer

IssuersOrVerifiers

IssuersWith

A map from Issuer to some value.

Schemas

StorageAccesses

Number of times storage entities were accessed.

TrustRegistryId

Unique identifier for the TrustRegistry.

TrustRegistryIdForParticipants

TrustRegistryIdSet

TrustRegistryInfo

TrustRegistryIssuerConfiguration

TrustRegistrySchemaId

Unique identifier for the TrustRegistry.

TrustRegistrySchemaMetadata

TrustRegistry schema metadata.

TrustRegistrySchemaVerifiers

Schema Verifiers.

TrustRegistryStoredParticipantInformation

Information about the TrustRegistry participant.

TrustRegistryStoredParticipants

Set of TrustRegistry participants.

TrustRegistryStoredSchemas

Set of schemas that belong to the TrustRegistry

UnboundedDelegatedIssuers

Unbounded delegated Issuers.

UnboundedIssuersWith

An unbounded map from Issuer to some value.

UnboundedSchemas

UnboundedTrustRegistryParticipantInformation

Information about the TrustRegistry participant.

UnboundedTrustRegistrySchemaMetadata

Unbounded TrustRegistry schema metadata.

UnboundedTrustRegistrySchemaVerifiers

Schema Verifiers.

UnboundedVerificationPrices

Prices of verifying a credential corresponding to the specific schema metadata per different currencies.

VerificationPrice

Price to verify a credential. Lowest denomination should be used.

VerificationPrices

Prices of verifying a credential corresponding to the specific schema metadata per different currencies.

Verifier

Trust registry Verifier’s DID.

VerifierSchemas

Set of schemas corresponding to a verifier

VerifierTrustRegistries

Set of trust registries corresponding to a verifier

Traits

HasSchemasAndRegistries

An entity that has references to schemas and registries.

Type Definitions

AggregatedTrustRegistrySchemaIssuers

DelegatedUpdate

MultiSchemaIdUpdate

RegistryIdUpdate

SchemaIdUpdate

TrustRegistrySchemaIssuers

Schema Issuers (Issuer => verification prices).

UnboundedDelegatedIssuersUpdate

UnboundedRegistrySchemaIssuers

UnboundedTrustRegistryParticipantsUpdate