Trait dock_core::util::action::Action

pub trait Action: Sized {
    type Target;

    // Required methods
    fn target(&self) -> Self::Target;
    fn len(&self) -> u32;

    // Provided methods
    fn is_empty(&self) -> bool { ... }
    fn modify<T, S, F, R, E>(self, f: F) -> Result<R, E>
       where F: FnOnce(Self, &mut S) -> Result<R, E>,
             <Self::Target as Associated<T>>::Value: TryInto<S>,
             S: Into<<Self::Target as Associated<T>>::Value>,
             E: From<ActionExecutionError> + From<NonceError>,
             Self::Target: StorageRef<T> { ... }
    fn view<T, S, F, R, E>(self, f: F) -> Result<R, E>
       where F: FnOnce(Self, S) -> Result<R, E>,
             <Self::Target as Associated<T>>::Value: TryInto<S>,
             S: Into<<Self::Target as Associated<T>>::Value>,
             E: From<ActionExecutionError> + From<NonceError>,
             Self::Target: StorageRef<T> { ... }
    fn modify_removable<T, S, F, R, E>(self, f: F) -> Result<R, E>
       where F: FnOnce(Self, &mut Option<S>) -> Result<R, E>,
             <Self::Target as Associated<T>>::Value: TryInto<S>,
             S: Into<<Self::Target as Associated<T>>::Value>,
             E: From<ActionExecutionError> + From<NonceError>,
             Self::Target: StorageRef<T> { ... }
    fn multi_signed<T, S, SI>(
        self,
        signatures: SI
    ) -> MultiSignedAction<T, Self, S, SI::IntoIter>
       where T: Types,
             S: Signature,
             SI: IntoIterator,
             SI::IntoIter: FusedIterator<Item = SignatureWithNonce<T::BlockNumber, S>> { ... }
}

Describes an action which can be performed on some Target.

Required Associated Types

type Target

Action target.

Required Methods

fn target(&self) -> Self::Target

Returns underlying action target.

fn len(&self) -> u32

Returns action unit length.

Provided Methods

fn is_empty(&self) -> bool

Returns true if the action unit count is equal to zero.

fn modify<T, S, F, R, E>(self, f: F) -> Result<R, E>where F: FnOnce(Self, &mut S) -> Result<R, E>, <Self::Target as Associated<T>>::Value: TryInto<S>, S: Into<<Self::Target as Associated<T>>::Value>, E: From<ActionExecutionError> + From<NonceError>, Self::Target: StorageRef<T>,

Calls supplied function accepting an action along with a mutable reference to the value associated with the target.

fn view<T, S, F, R, E>(self, f: F) -> Result<R, E>where F: FnOnce(Self, S) -> Result<R, E>, <Self::Target as Associated<T>>::Value: TryInto<S>, S: Into<<Self::Target as Associated<T>>::Value>, E: From<ActionExecutionError> + From<NonceError>, Self::Target: StorageRef<T>,

Calls supplied function accepting an action along with a value associated with the target.

fn modify_removable<T, S, F, R, E>(self, f: F) -> Result<R, E>where F: FnOnce(Self, &mut Option<S>) -> Result<R, E>, <Self::Target as Associated<T>>::Value: TryInto<S>, S: Into<<Self::Target as Associated<T>>::Value>, E: From<ActionExecutionError> + From<NonceError>, Self::Target: StorageRef<T>,

Calls supplied function accepting an action along with a mutable reference to the option possibly containing a value associated with the target. Modifying supplied Option<_> to None will lead to the value removal.

fn multi_signed<T, S, SI>( self, signatures: SI ) -> MultiSignedAction<T, Self, S, SI::IntoIter>where T: Types, S: Signature, SI: IntoIterator, SI::IntoIter: FusedIterator<Item = SignatureWithNonce<T::BlockNumber, S>>,

Combines underlying action with the provided signatures.