Trait dock_core::util::action::ActionWithNonce

pub trait ActionWithNonce<T: Types>: Action {
    // Required method
    fn nonce(&self) -> T::BlockNumber;

    // Provided methods
    fn execute_and_increase_nonce<F, S, R, E>(self, f: F) -> Result<R, E>
       where F: FnOnce(Self, &mut Option<S>) -> Result<R, E>,
             E: From<ActionExecutionError> + From<NonceError>,
             Self::Target: StorageRef<T>,
             <Self::Target as Associated<T>>::Value: TryInto<WithNonce<T, S>>,
             WithNonce<T, S>: Into<<Self::Target as Associated<T>>::Value> { ... }
    fn execute_without_increasing_nonce<F, R, S, E>(self, f: F) -> Result<R, E>
       where F: FnOnce(Self, &mut Option<S>) -> Result<R, E>,
             E: From<ActionExecutionError>,
             WithNonce<T, S>: TryFrom<<Self::Target as Associated<T>>::Value> + Into<<Self::Target as Associated<T>>::Value>,
             Self::Target: StorageRef<T>,
             <Self::Target as Associated<T>>::Value: TryInto<WithNonce<T, S>> { ... }
    fn signed<S>(self, signature: S) -> SignedActionWithNonce<T, Self, S> { ... }
    fn signed_with_combined_target<S, F, Ta>(
        self,
        signature: S,
        build_target: F
    ) -> Result<SignedActionWithNonce<T, ActionWithNonceWrapper<T, Self, Ta>, S>, InvalidSigner>
       where S: Signature,
             F: FnOnce(Self::Target, S::Signer) -> Ta,
             Ta: Clone { ... }
    fn signed_with_signer_target<S>(
        self,
        signature: S
    ) -> Result<SignedActionWithNonce<T, ActionWithNonceWrapper<T, Self, S::Signer>, S>, InvalidSigner>
       where S: Signature { ... }
}

Describes an action with nonce which can be performed on some Target

Required Methods

fn nonce(&self) -> T::BlockNumber

Returns action’s nonce.

Provided Methods

fn execute_and_increase_nonce<F, S, R, E>(self, f: F) -> Result<R, E>where F: FnOnce(Self, &mut Option<S>) -> Result<R, E>, E: From<ActionExecutionError> + From<NonceError>, Self::Target: StorageRef<T>, <Self::Target as Associated<T>>::Value: TryInto<WithNonce<T, S>>, WithNonce<T, S>: Into<<Self::Target as Associated<T>>::Value>,

Executes an action providing a mutable reference to the option containing a value associated with the target. In case of a successful result, the nonce will be increased.

fn execute_without_increasing_nonce<F, R, S, E>(self, f: F) -> Result<R, E>where F: FnOnce(Self, &mut Option<S>) -> Result<R, E>, E: From<ActionExecutionError>, WithNonce<T, S>: TryFrom<<Self::Target as Associated<T>>::Value> + Into<<Self::Target as Associated<T>>::Value>, Self::Target: StorageRef<T>, <Self::Target as Associated<T>>::Value: TryInto<WithNonce<T, S>>,

Executes an action providing a mutable reference to the value associated with the target. Even in case of a successful result, the nonce won’t be increased.

fn signed<S>(self, signature: S) -> SignedActionWithNonce<T, Self, S>

Combines underlying action with the provided signature.

fn signed_with_combined_target<S, F, Ta>( self, signature: S, build_target: F ) -> Result<SignedActionWithNonce<T, ActionWithNonceWrapper<T, Self, Ta>, S>, InvalidSigner>where S: Signature, F: FnOnce(Self::Target, S::Signer) -> Ta, Ta: Clone,

Wraps underlying action into an action targeting signer then combines result with the provided signature.

fn signed_with_signer_target<S>( self, signature: S ) -> Result<SignedActionWithNonce<T, ActionWithNonceWrapper<T, Self, S::Signer>, S>, InvalidSigner>where S: Signature,

Wraps underlying action into an action targeting signer then combines result with the provided signature.