Trait dock_core::util::action::StorageRef

pub trait StorageRef<T>: Sized {
    type Value;

    // Required methods
    fn try_mutate_associated<F, R, E>(self, f: F) -> Result<R, E>
       where F: FnOnce(&mut Option<Self::Value>) -> Result<R, E>;
    fn view_associated<F, R>(self, f: F) -> R
       where F: FnOnce(Option<Self::Value>) -> R;
}

Allows to view and mutate a value associated with Self.

Required Associated Types

type Value

Some value type associated with Self.

Required Methods

fn try_mutate_associated<F, R, E>(self, f: F) -> Result<R, E>where F: FnOnce(&mut Option<Self::Value>) -> Result<R, E>,

Attempts to mutate a value associated with Self. If the value under the option is taken, the associated value will be removed. All updates will be applied only in case of a successful result.

fn view_associated<F, R>(self, f: F) -> Rwhere F: FnOnce(Option<Self::Value>) -> R,

Calls provided function with an associated value as an argument.