Trait dock_core::util::batch_update::CanUpdateKeyed

pub trait CanUpdateKeyed<Entity>where
    Entity: Deref,
    Entity::Target: KeyValue,{
    // Provided method
    fn can_update_keyed<U: KeyedUpdate<Entity>>(
        &self,
        _entity: &Entity,
        _update: &U
    ) -> bool { ... }
}

Checks whether an actor can update an entity over some keys.

Provided Methods

fn can_update_keyed<U: KeyedUpdate<Entity>>( &self, _entity: &Entity, _update: &U ) -> bool

Checks whether the underlying keyed update can be applied, i.e. all associated updates are valid.